refactor(test): extract Phase 1 wire harness into tests/support/wire
Pure relocation: tests/support/{mod.rs,wire/{mod.rs,harness.rs,schema.rs}}
own the Harness, constants, and MCP spec schema validators that
mcp_wire_conformance.rs used to declare inline. Test bodies and
assertions are unchanged; the file imports the moved items via
#[path = "support/mod.rs"] mod support; use support::wire::…;.

Sets up the shared driver Phase 3 (#265) builds on.
