+++ crates/rimap-server/tests/mcp_wire_conformance.rs
@@ -1,301 +1,21 @@
 //! MCP wire-shape conformance harness (issue #263, Phase 1).
 //!
-//! Spawns the production `rusty-imap-mcp` binary with a tempdir
-//! config that has zero accounts, drives a deterministic JSON-RPC
-//! sequence over its stdio, and validates every response against
-//! the vendored MCP spec schemas under
-//! `tests/fixtures/mcp-spec/2025-11-25/`.
-//!
-//! Permanent regression net for two real wire-shape bugs:
-//!  - #261: `initialize.result.capabilities` was serialized as `{}`.
-//!  - `fix/tool-input-schema-object-type`: tools advertised
-//!    `inputSchema: {}` with no `type` field.
+//! Drives the production `rusty-imap-mcp` binary over stdio with a
+//! zero-account config and validates every response against the
+//! vendored MCP spec schemas. See
+//! `docs/superpowers/specs/2026-05-12-mcp-wire-conformance-design.md`.
 
 #![expect(clippy::expect_used, reason = "integration tests")]
-#![expect(clippy::panic, reason = "test assertions render diagnostics")]
 
-use std::collections::HashMap;
-use std::process::Stdio;
-use std::sync::{Arc, Mutex, OnceLock};
+#[path = "support/mod.rs"]
+mod support;
+
 use std::time::Duration;
 
-use assert_cmd::cargo::cargo_bin;
 use rmcp::model::ProtocolVersion;
-use serde_json::{Value, json};
-use tempfile::TempDir;
-use tokio::io::{AsyncBufReadExt, AsyncWriteExt, BufReader};
-use tokio::process::{Child, ChildStdin, ChildStdout, Command};
-use tokio::time::timeout;
+use serde_json::json;
 
-/// MCP protocol version pinned by this harness. Matches the
-/// directory under `tests/fixtures/mcp-spec/` and the `LATEST` value
-/// in `rmcp 1.5`. Update both when bumping.
-const PINNED_PROTOCOL_VERSION: &str = "2025-11-25";
-
-/// Vendored MCP spec schema, compiled in at build time so tests run
-/// hermetically (no network, no filesystem dependency beyond the
-/// crate source).
-const MCP_SCHEMA_JSON: &str = include_str!("fixtures/mcp-spec/2025-11-25/schema.json");
-
-const REQUEST_TIMEOUT: Duration = Duration::from_secs(2);
-// Under `cargo nextest run` with the full workspace suite (~1100 tests
-// in parallel), the EOF-to-exit slice for `wire_clean_eof_shutdown_exits_zero`
-// can exceed a tight 1 s budget on CPU-contended runners. 5 s remains
-// tight enough to fail-fast on a real hang while absorbing scheduling
-// jitter when other tests are spawning binaries / parsers concurrently.
-const SHUTDOWN_TIMEOUT: Duration = Duration::from_secs(5);
-
-/// Owns the spawned child plus its piped stdio.
-struct Harness {
-    child: Child,
-    stdin: ChildStdin,
-    stdout: BufReader<ChildStdout>,
-    next_id: u64,
-    // Hold the tempdir until the harness drops so the audit log path
-    // remains valid for the lifetime of the spawned process.
-    _tempdir: TempDir,
-}
-
-impl Harness {
-    /// Spawn the binary with a zero-account tempdir config.
-    #[expect(
-        clippy::unused_async,
-        reason = "harness API is uniformly async so tests await every constructor"
-    )]
-    async fn spawn() -> Self {
-        let tempdir = TempDir::new().expect("tempdir");
-        let config_path = tempdir.path().join("config.toml");
-        let audit_path = tempdir.path().join("audit.jsonl");
-        let allowed_base = tempdir.path();
-
-        // Multi-account format with zero accounts. Task 1 lifted the
-        // empty-accounts validator gate, so this is the canonical
-        // zero-account shape the loader accepts.
-        let config = format!(
-            r#"
-accounts = []
-
-[audit]
-path = "{}"
-allowed_base_dir = "{}"
-"#,
-            audit_path.display(),
-            allowed_base.display(),
-        );
-        std::fs::write(&config_path, config).expect("write config");
-
-        let mut cmd = Command::new(cargo_bin("rusty-imap-mcp"));
-        cmd.arg("--config")
-            .arg(&config_path)
-            // Production rejects `accounts = []`. The harness opts in
-            // to infrastructure-only boot via this test-support flag
-            // (Codex adversarial review on PR #270).
-            .arg("--allow-empty-accounts")
-            .stdin(Stdio::piped())
-            .stdout(Stdio::piped())
-            .stderr(Stdio::null())
-            .kill_on_drop(true);
-
-        let mut child = cmd.spawn().expect("spawn rusty-imap-mcp binary");
-
-        let stdin = child.stdin.take().expect("stdin");
-        let stdout = BufReader::new(child.stdout.take().expect("stdout"));
-
-        Self {
-            child,
-            stdin,
-            stdout,
-            next_id: 0,
-            _tempdir: tempdir,
-        }
-    }
-
-    /// Send a JSON-RPC request and return the parsed response value.
-    /// Panics on timeout, EOF before a response arrives, or non-JSON output.
-    async fn request(&mut self, method: &str, params: Value) -> Value {
-        self.next_id += 1;
-        let id = self.next_id;
-        let envelope = json!({
-            "jsonrpc": "2.0",
-            "id": id,
-            "method": method,
-            "params": params,
-        });
-        let line = format!("{envelope}\n");
-        self.stdin
-            .write_all(line.as_bytes())
-            .await
-            .expect("write request");
-        self.stdin.flush().await.expect("flush request");
-
-        let mut buf = String::new();
-        let read = timeout(REQUEST_TIMEOUT, self.stdout.read_line(&mut buf))
-            .await
-            .expect("response within timeout")
-            .expect("read response");
-        assert!(read > 0, "stdout closed before responding to {method}");
-        let response: Value = serde_json::from_str(buf.trim_end()).expect("parse response JSON");
-        assert_eq!(response["id"], json!(id), "response id must match request");
-        assert_envelope_valid(&response);
-        response
-    }
-
-    /// Send a JSON-RPC notification (no `id`, no response expected).
-    async fn notify(&mut self, method: &str, params: Value) {
-        let envelope = json!({
-            "jsonrpc": "2.0",
-            "method": method,
-            "params": params,
-        });
-        let line = format!("{envelope}\n");
-        self.stdin
-            .write_all(line.as_bytes())
-            .await
-            .expect("write notification");
-        self.stdin.flush().await.expect("flush notification");
-    }
-
-    /// Assert no bytes arrive on stdout for the given duration.
-    async fn assert_no_response_within(&mut self, dur: Duration) {
-        let mut buf = String::new();
-        match timeout(dur, self.stdout.read_line(&mut buf)).await {
-            Err(_) => {} // timeout → no response, as expected
-            Ok(Ok(0)) => panic!("stdout closed unexpectedly"),
-            Ok(Ok(_)) => panic!("expected no response within {dur:?}, got: {buf:?}"),
-            Ok(Err(e)) => panic!("read error: {e}"),
-        }
-    }
-
-    /// Send an MCP `initialize` request with the pinned protocol
-    /// version and return the response.
-    async fn initialize_handshake(&mut self) -> Value {
-        self.request(
-            "initialize",
-            json!({
-                "protocolVersion": ProtocolVersion::LATEST.as_str(),
-                "capabilities": {},
-                "clientInfo": {
-                    "name": "rusty-imap-mcp-conformance-harness",
-                    "version": env!("CARGO_PKG_VERSION"),
-                },
-            }),
-        )
-        .await
-    }
-
-    /// Send `notifications/initialized` after the handshake.
-    async fn send_initialized(&mut self) {
-        self.notify("notifications/initialized", json!({})).await;
-    }
-
-    /// Close stdin, await the child, and return its exit status.
-    async fn shutdown_and_wait(mut self) -> std::process::ExitStatus {
-        drop(self.stdin);
-        timeout(SHUTDOWN_TIMEOUT, self.child.wait())
-            .await
-            .expect("clean exit within timeout")
-            .expect("wait")
-    }
-}
-
-/// Compile (once) and return a validator for a top-level definition in
-/// the vendored MCP schema. `fragment` is the key under
-/// `definitions` / `$defs` (e.g. `"InitializeResult"`). Returns an
-/// `Arc` so multiple parallel tests can share the compiled validator
-/// without lifetime gymnastics.
-fn validator_for(fragment: &'static str) -> Arc<jsonschema::Validator> {
-    // All function-scoped items declared up front so cache and parsed
-    // schema lifetimes are visible from the top of the body
-    // (clippy::items_after_statements).
-    type Cache = Mutex<HashMap<&'static str, Arc<jsonschema::Validator>>>;
-    static PARSED: OnceLock<(Value, &'static str)> = OnceLock::new();
-    static CACHE: OnceLock<Cache> = OnceLock::new();
-
-    // Parse the vendored schema exactly once and detect the
-    // definitions key (`$defs` for Draft 2020-12, `definitions` for
-    // older dialects). The full Value and the detected key are
-    // immutable for the lifetime of the test process.
-    let parsed = PARSED.get_or_init(|| {
-        let full: Value = serde_json::from_str(MCP_SCHEMA_JSON).expect("parse vendored MCP schema");
-        let defs_key = if full.get("$defs").is_some() {
-            "$defs"
-        } else {
-            "definitions"
-        };
-        (full, defs_key)
-    });
-    let full = &parsed.0;
-    let defs_key: &'static str = parsed.1;
-
-    let cache = CACHE.get_or_init(|| Mutex::new(HashMap::new()));
-
-    // Fast path: read under the lock, return if cached.
-    {
-        let guard = cache.lock().expect("validator cache mutex poisoned");
-        if let Some(v) = guard.get(fragment) {
-            return Arc::clone(v);
-        }
-    }
-
-    // Slow path: compile the fragment validator WITHOUT holding the
-    // cache lock, so parallel tests compiling different fragments
-    // don't serialize.
-    let wrapper = json!({
-        "$ref": format!("#/{defs_key}/{fragment}"),
-        defs_key: full.get(defs_key).cloned().unwrap_or(json!({})),
-    });
-    let new_validator =
-        Arc::new(jsonschema::validator_for(&wrapper).expect("compile fragment validator"));
-
-    // Insert. If a concurrent thread already inserted while we were
-    // compiling, prefer the existing entry to keep the Arc stable.
-    let mut guard = cache.lock().expect("validator cache mutex poisoned");
-    let entry = guard
-        .entry(fragment)
-        .or_insert_with(|| Arc::clone(&new_validator));
-    Arc::clone(entry)
-}
-
-/// Validate a value against a vendored fragment, panicking with the
-/// list of errors on failure.
-fn assert_valid(value: &Value, fragment: &'static str) {
-    let v = validator_for(fragment);
-    if !v.is_valid(value) {
-        let errors: Vec<String> = v.iter_errors(value).map(|e| e.to_string()).collect();
-        panic!(
-            "schema validation failed for fragment {fragment}:\n  {}",
-            errors.join("\n  ")
-        );
-    }
-}
-
-/// Validate the FULL JSON-RPC envelope returned by `Harness::request`.
-/// Success responses validate against `JSONRPCResultResponse`; error
-/// responses validate against `JSONRPCErrorResponse`. Asserts the
-/// `jsonrpc` version field on both paths. Codex adversarial review
-/// finding #2 (PR #270): the previous negative-path tests checked only
-/// `code` and `message` and would have missed a regression that
-/// stripped `jsonrpc` or otherwise mangled the envelope.
-fn assert_envelope_valid(response: &Value) {
-    assert_eq!(
-        response["jsonrpc"],
-        json!("2.0"),
-        "envelope must declare jsonrpc=\"2.0\"; got {response}",
-    );
-
-    let has_result = response.get("result").is_some();
-    let has_error = response.get("error").is_some();
-    match (has_result, has_error) {
-        (true, false) => assert_valid(response, "JSONRPCResultResponse"),
-        (false, true) => assert_valid(response, "JSONRPCErrorResponse"),
-        (true, true) => {
-            panic!("envelope must not contain both `result` and `error`; got {response}",)
-        }
-        (false, false) => {
-            panic!("envelope must contain either `result` or `error`; got {response}",)
-        }
-    }
-}
+use support::wire::{Harness, PINNED_PROTOCOL_VERSION, assert_valid};
 
 #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
 async fn wire_smoke_initialize_returns_valid_envelope() {
@@ -342,7 +62,7 @@ async fn wire_protocol_version_negotiation_matches_vendored_schema() {
     // Three-way drift check (Codex adversarial review finding #1):
     //
     //   1. rmcp::ProtocolVersion::LATEST.as_str()
-    //   2. PINNED_PROTOCOL_VERSION (the constant in this file)
+    //   2. PINNED_PROTOCOL_VERSION (constant in tests/support/wire/harness.rs)
     //   3. crates/rimap-server/tests/fixtures/mcp-spec/<version>/
     //
     // All three MUST agree. If any one drifts (rmcp bumps LATEST, the
@@ -358,7 +78,7 @@ async fn wire_protocol_version_negotiation_matches_vendored_schema() {
          PINNED_PROTOCOL_VERSION ({PINNED_PROTOCOL_VERSION}). Run \
          `scripts/refresh-mcp-spec.sh {rmcp_latest}` to vendor the new \
          schema, update PINNED_PROTOCOL_VERSION + MCP_SCHEMA_JSON in \
-         this file, and update the README under \
+         tests/support/wire/harness.rs, and update the README under \
          tests/fixtures/mcp-spec/.",
     );
